Updated mainly the comment and ticket resources, and fixed some minor issues.

+++ code/backend/jvm/src/main/kotlin/pt/isel/ps/project/model/person/PersonEntity.kt
@@ -29,8 +29,8 @@ data class CreatePersonEntity(
     val email: String,
     val password: String,
     val role: String,
-    val company: Int,
-    val skill: Int? = null,
+    val company: Long,
+    val skill: Long? = null,
 )
 
 data class UpdatePersonEntity(
